+++ src/components/manage/Blocks/Text/Edit.jsx
@@ -45,6 +45,7 @@ class Edit extends Component {
     selected: PropTypes.bool.isRequired,
     block: PropTypes.string.isRequired,
     onAddBlock: PropTypes.func.isRequired,
+    onAddBlockBefore: PropTypes.func.isRequired,
     onChangeBlock: PropTypes.func.isRequired,
     onDeleteBlock: PropTypes.func.isRequired,
     onMutateBlock: PropTypes.func.isRequired,
@@ -321,11 +322,11 @@ class Edit extends Component {
           )}
         {this.state.addNewBlockOpened && (
           <BlockChooser
-            onMutateBlock={(id, value) => {
+            onAddBlockBefore={(id, value) => {
               this.setState((state) => ({
                 addNewBlockOpened: !state.addNewBlockOpened,
               }));
-              this.props.onSelectBlock(this.props.onMutateBlock(id, value));
+              this.props.onSelectBlock(this.props.onAddBlockBefore(id, value));
             }}
             currentBlock={this.props.block}
             allowedBlocks={this.props.allowedBlocks}